move `normalize_datetime` to a separate module
