+++ ceos_alos2/volume_directory/metadata.py
@@ -1,14 +1,9 @@
-import datetime as dt
-
 from tlz.functoolz import curry, pipe
 
 from ceos_alos2.dicttoolz import apply_to_items, dissoc
 from ceos_alos2.hierarchy import Group
+from ceos_alos2.transformers import normalize_datetime
 from ceos_alos2.utils import remove_nesting_layer, rename
-
-
-def normalize_datetime(string):
-    return dt.datetime.strptime(string, "%Y%m%d%H%M%S%f").isoformat()
 
 
 def transform_volume_descriptor(mapping):
